fix: fixed bug with feature flag handling
